# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Customizer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotProfile {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  bannerUrl: string | null;
  /** @nullable */
  bannerColor: string | null;
}

export interface AvatarInput {
  /** Base64-encoded data URI of the image (e.g. data:image/png;base64,...) */
  imageDataUri: string;
}

export interface BannerInput {
  /** Base64-encoded data URI of the image (e.g. data:image/png;base64,...) */
  imageDataUri: string;
}

export interface ErrorResponse {
  error: string;
}