# lib/api-zod/src/generated/types/bannerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Customizer API
 * OpenAPI spec version: 0.1.0
 */

export interface BannerInput {
  /** Base64-encoded data URI of the image (e.g. data:image/png;base64,...) */
  imageDataUri: string;
}