# lib/api-zod/src/generated/types/botProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Customizer API
 * OpenAPI spec version: 0.1.0
 */

export interface BotProfile {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  bannerUrl: string | null;
  /** @nullable */
  bannerColor: string | null;
}